+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.0.2: 2024-06-23
+
+* Declare PHP 8.3 support in phpcs.xml
+* Add autoupdate from GitHub as suggested by @cagrimmett, Fixes #2
+* Better plugin description
+
 ### 1.0.1: 2023-02-25
 
 * Add Koko Analytics support